backend/studygpt/api/serializers.py
from rest_framework import serializers
from .models import Question

class QuestionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Question
        fields = ['id', 'user_question', 'subject', 'language', 'answer', 'created_at']