backend/requirements.txt
django==4.2.11
djangorestframework==3.15.1
requests==2.31.0
python-decouple==3.8
django-cors-headers==4.3.1